transaction history and receipt done (#46)
* added transaction history (#22)

* added dashboard

* added AppDrawer component

* added transaction history

---------


* Database functions (#24)

* database functions

* initial function testing

* partial exception functions

---------


* image test working init on product upload

* still testing product image thingy

* save image post final

* post backend integration for products

* dropdown thing

* remove description prdocut

* needs refactor type shit and edit is gucci prlly

* image on edit products ting ug tung duplicate thing problem

* edit products image is now updatable

* transactions tentative

* transactions and shit

* transaction history done

---------

+++ lib/models/OrdersModel.dart
@@ -18,7 +18,7 @@ class Orders {
       id: orderDoc,
       product_name: data['product_id'],
       productImage: data['productImage'],
-      total_price: data['total_price'],
+      total_price: data['total_price'].toDouble(),
       quantity: data['quantity'],
     );
   }
